[#22] fix: 트랜잭션 경계 수정
